[#29]Feat: 차트조회 API 엔드포인트 변경 -> GameController
