+++ src/main/java/capstone/SportyUp/SportyUp_Server/converter/GameConverter.java
@@ -18,4 +18,13 @@ public class GameConverter {
                 .highlightUrl(game.getHighlightUrl())
                 .build();
     }
+
+    public static GameResponseDTO.ChartDTO toChartDTO(Long gameCount, Double averageScore, Integer highScore, Integer lowScore){
+        return GameResponseDTO.ChartDTO.builder()
+                .gameCount(gameCount)
+                .averageScore(averageScore)
+                .highScore(highScore)
+                .lowScore(lowScore)
+                .build();
+    }
 }
