+++ src/main/java/capstone/SportyUp/SportyUp_Server/service/GameService/GameQueryService.java
@@ -1,0 +1,7 @@
+package capstone.SportyUp.SportyUp_Server.service.GameService;
+
+import capstone.SportyUp.SportyUp_Server.web.DTO.GameDTO.GameResponseDTO;
+
+public interface GameQueryService {
+    public GameResponseDTO.ChartDTO getChart(Long userId, Long sportsId);
+}
