+++ src/main/java/capstone/SportyUp/SportyUp_Server/web/controller/GameController.java
@@ -3,6 +3,7 @@ package capstone.SportyUp.SportyUp_Server.web.controller;
 
 import capstone.SportyUp.SportyUp_Server.apiPayload.ApiResponse;
 import capstone.SportyUp.SportyUp_Server.service.GameService.GameCommandService;
+import capstone.SportyUp.SportyUp_Server.service.GameService.GameQueryService;
 import capstone.SportyUp.SportyUp_Server.web.DTO.GameDTO.GameRequestDTO;
 import capstone.SportyUp.SportyUp_Server.web.DTO.GameDTO.GameResponseDTO;
 import capstone.SportyUp.SportyUp_Server.web.controller.specification.GameSpecification;
@@ -18,6 +19,7 @@ import java.time.LocalDate;
 public class GameController implements GameSpecification {
 
     private final GameCommandService gameCommandService;
+    private final GameQueryService gameQueryService;
 
     @Override
     public ApiResponse<GameResponseDTO.GameInfoListDTO> getGameList(LocalDate date, String sports) {
@@ -45,5 +47,11 @@ public class GameController implements GameSpecification {
 
 
         return ApiResponse.onSuccess(gameCommandService.endGame(gameId, request));
+    }
+
+    @Override
+    public ApiResponse<GameResponseDTO.ChartDTO> getChart(Long userId, Long sportsId) {
+
+        return ApiResponse.onSuccess(gameQueryService.getChart(userId, sportsId));
     }
 }
