+++ src/main/java/capstone/SportyUp/SportyUp_Server/web/controller/specification/GameSpecification.java
@@ -48,4 +48,10 @@ public interface GameSpecification {
     })
     ApiResponse<GameResponseDTO.EndResultDTO> endGame(@PathVariable Long gameId, @RequestBody GameRequestDTO.EndDTO request);
 
+    @GetMapping("/chart/{sportsId}")
+    @Operation(summary = "분석 페이지 차트 조회 API", description = "차트에 필요한 데이터를 조회하는 API입니다. 어떤 종목에 대한 차트인지 구분 필요")
+    @ApiResponses({
+            @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "COMMON200", description = "OK, 성공"),
+    })
+    ApiResponse<GameResponseDTO.ChartDTO> getChart(@RequestParam Long userId, @PathVariable Long sportsId);
 }
